pyproject.toml
[tool.poetry]
name = "phai-core-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
python-dotenv = "^1.0.1"
gemini-functions-agent = {path = "packages/gemini-functions-agent", develop = true}
pydantic = "1.10.13"
rag-supabase = {path = "packages/rag-supabase", develop = true}
google-cloud-secret-manager = "^2.19.0"
langgraph = "^0.0.48"
langchain-openai = "^0.0.5"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
